Entity API: security upgrade

+++ ci-scripts/pre_cache.sh
@@ -43,7 +43,10 @@ if [[ "$PREVIOUS_NPM_HASH" == "$CURRENT_NPM_HASH" && -d "$TRAVIS_CACHE_DIR"/node
   cp -r "$TRAVIS_CACHE_DIR"/node_modules "$ROOT_DIR"/client
 else
   echo "NPM build hash does not match, purging cache ($PREVIOUS_NPM_HASH <> $CURRENT_NPM_HASH)"
-  rm -rf "$TRAVIS_CACHE_DIR"/node_modules
+  if [ -d "$TRAVIS_CACHE_DIR"/node_modules ]; then
+    chmod -R +w "$TRAVIS_CACHE_DIR"/node_modules
+    rm -rf "$TRAVIS_CACHE_DIR"/node_modules
+  fi
 fi
 
 if [[ "$PREVIOUS_DRUPAL_HASH" == "$CURRENT_DRUPAL_HASH" && -d "$TRAVIS_CACHE_DIR"/www ]]; then
@@ -52,6 +55,9 @@ if [[ "$PREVIOUS_DRUPAL_HASH" == "$CURRENT_DRUPAL_HASH" && -d "$TRAVIS_CACHE_DIR
   cp -r "$TRAVIS_CACHE_DIR"/www_ignored/* "$ROOT_DIR"/
 else
   echo "Drupal build hash does not match, purging cache ($PREVIOUS_DRUPAL_HASH <> $CURRENT_DRUPAL_HASH)"
-  rm -rf "$TRAVIS_CACHE_DIR"/www
+  if [ -d "$TRAVIS_CACHE_DIR"/www ]; then
+    chmod -R +w "$TRAVIS_CACHE_DIR"/www
+    rm -rf "$TRAVIS_CACHE_DIR"/www
+  fi
 fi
 date
